+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix crash on import with `rics>=4.1.0`.
+
 ## [0.10.1] - 2024-04-17
 
 ### Fixed